fix(api): correct max_score_milli decoding and Redis-aware competition test
- contest_problems.max_score_milli is INT4: decode it as i32 in the
  result context (the two judge-result integration tests failed on the
  i64/i32 column mismatch).
- The competition workstation-session integration test now connects to
  the integration Redis (PROJECT_BALLOON_TEST_REDIS_URL) since session
  creation requires it, and skips when the variable is absent.

+++ apps/api/src/features/judge_dispatch/result_processor.rs
@@ -58,7 +58,7 @@ struct ResultContext {
     start_at: Option<OffsetDateTime>,
     scoring_icpc: bool,
     aggregation_best: bool,
-    max_score_milli: i64,
+    max_score_milli: i32,
 }
 
 impl JudgeResultProcessor {
@@ -339,7 +339,7 @@ impl JudgeResultProcessor {
             scoring_icpc: context.scoring_icpc,
             aggregation_best: context.aggregation_best,
             score_milli,
-            max_score_milli: context.max_score_milli,
+            max_score_milli: i64::from(context.max_score_milli),
         };
         if let Err(error) = projection.apply_judgement(&event).await {
             tracing::warn!(
